requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper

# Audio processing dependency (Whisper requires ffmpeg installed on system)
# Note! ffmpeg is NOT a pip dependency and should be installed seperatly (instructions in README.md)